moving linkml datastructure pieces out of dataharmonizer.js
