+++ lib/AppContext.js
@@ -7,12 +7,14 @@ import {
   Template,
   getTemplatePathInScope,
 } from '../lib/utils/templates';
-import { wait, isEmpty } from '../lib/utils/general';
+
+import { wait, isEmpty, pascalToLowerWithSpaces } from '../lib/utils/general';
 import { invert, removeNumericKeys, consolidate } from '../lib/utils/objects';
 import { createDataHarmonizerContainer, createDataHarmonizerTab } from '../web';
 import { getExportFormats } from 'schemas';
 import {range as arrayRange} from 'lodash';
-
+import {deepMerge} from '../lib/utils/objects';
+import Validator from './Validator';
 
 class AppConfig {
   constructor(template_path = null) {
@@ -21,13 +23,14 @@ class AppConfig {
   }
 }
 export default class AppContext {
-  schema_tree = {};
+  //schema_tree = {};
   dhs = {};
   current_data_harmonizer_name = null;
   currentSelection = null;
   template = null;
   export_formats = {};
   oneToManyAppContext = null;
+  field_settings = {};
 
   constructor(appConfig = new AppConfig(getTemplatePathInScope())) {
     this.appConfig = appConfig;
@@ -57,38 +60,571 @@ export default class AppContext {
     });
   }
 
-  async reload(
-    template_path,
-    overrides = { locale: null, forced_schema: null }
-  ) {
-    const isSameTemplatePath = template_path === this.appConfig.template_path;
-    const isUploadedTemplate = template_path.startsWith('local');
-    const isLocaleChange = overrides.locale !== null;
+  /*
+   * Return initialized, rendered dataHarmonizer instances rendered in order
+   * of their appearance as classes in schema. If a template name (class name)
+   * is provided, only include that one and any dependent tables in its schema.
+   * Each class has one or more slots (attributes) which each contain a 
+   * foreign_key which determines parent-child relationships.
+   *
+   * Uses this.relations to construct list of main DH class and its dependant 
+   * classes.
+   *
+   * Processing each class into DataHarmonizer tables:
+   * - One DH instance per class: this.dhs[template_name]
+   *   - In rendered DH view, filter its table rows by the foreign key slots 
+   *     featured in that class.
+   * - Navigation: One tab per DH instance.
+   */
+  async reload(template_path, locale = null, forced_schema= null) {
 
-    // Handle local template refresh case
-    if (isSameTemplatePath && isUploadedTemplate) {
-      return this.setupDataHarmonizers({dataharmonizers: this.dhs, forced_schema: this.template?.default?.schema});
+    this.clearInterface();
+    this.appConfig = new AppConfig(template_path);
+    const [schema_path, template_name] = this.appConfig.template_path.split('/');
+
+    //alert('template:' + schema_path + ':' + template_path + ':' + this.appConfig.template_path)
+    this.template = await Template.create(schema_path, {forced_schema});
+    if (locale !== null) {
+      this.template.updateLocale(locale);
     }
 
-    // Handle changes in schema template path or locale
-    if (!isSameTemplatePath || isLocaleChange) {
-      this.appConfig = new AppConfig(template_path);
-      return await this.setupDataHarmonizers({locale: overrides.locale});
+    // Empty case will occur when template_path doesn't correspond to a built template
+    this.export_formats = forced_schema ? {} : await getExportFormats(schema_path);
+
+    const schema =
+      locale !== null
+        ? this.template.localized.schema
+        : this.template.default.schema;
+
+    // .relations is a 2 pass construction. 1st pass, create parent/child
+    // relations. 2nd pass: create a carefully organized MAP 
+    // {[class_name]: dh...} which functions can process in left-to-right order
+    // to account for cascading actions on primary/foreign key relations.
+    this.relations = this.crudGetRelations(schema);
+    for (const class_name in schema.classes) {
+      if (class_name != 'Container' && class_name != 'dh_interface') {
+        this.relations[class_name].dependents = this.crudGetDependents(class_name);
+      }
     }
 
-    // Handle same template path without uploaded template
-    if (isSameTemplatePath) {
-      return this.setupDataHarmonizers({dataharmonizers: this.dhs});
-    }
-
-    // Handle forced schema case
-    if (overrides.forced_schema !== null) {
-      this.appConfig = new AppConfig(template_path);
-      return this.setupDataHarmonizers({forced_schema: overrides.forced_schema});
-    }
+    // Merges any existing dataharmonizer instances with the ones newly created.
+    this.dhs = this.makeDHsFromRelations(schema, template_name);
 
     // Default case: if no significant changes detected
     return this;
+  }
+
+  /**
+   * Creates Data Harmonizers from the schema tree.
+   * FUTURE: Make this a 2 pass thing.  First pass creates every dataharmonizer
+   * data structure; 2nd pass initializes handsontable part.
+   * @param {Object} schema_tree - The schema tree from which to create Data Harmonizers.
+   * @returns {Object} An object mapping the class names to their respective Data Harmonizer objects.
+   */
+  makeDHsFromRelations(schema, template_name) {
+    let data_harmonizers = {};
+
+    Object.entries(schema.classes)
+      // Container class is only used in input and output file coordination.
+      .filter(([class_name]) => class_name == template_name || this.relations?.[template_name]?.dependents?.has(class_name))
+      .forEach((obj, index) => {
+        if (obj.length > 0) {
+
+          const [class_name, class_obj] = obj;
+
+          // If it shares a key with another class which is its parent, this DH must be a child
+          const is_child = this.crudGetParents(class_name);
+
+          const dhId = `data-harmonizer-grid-${index}`;
+          const tab_label = class_obj.title ? class_obj.title : class_name;
+          const dhTab = createDataHarmonizerTab(dhId, tab_label, index === 0);
+          dhTab.addEventListener('click', () => {
+            $(document).trigger('dhTabChange', {
+              specName: class_name,
+            });
+          });
+
+
+          // Each selected DataHarmonizer is rendered here. 
+          // NOTE: this may be running twice?.
+          // in 1-M, different DataHarmonizers have fewer rows to start with
+          // Child tables are displayed differently so override the default
+          // HoT settings.
+          const dhSubroot = createDataHarmonizerContainer(dhId, index === 0);
+          const handsOnDH = new DataHarmonizer(dhSubroot, this, {
+            loadingScreenRoot: document.body,
+            template_name: class_name,
+            schema: schema, // assign during constructor so validator can init on it.
+            hot_override_settings: {
+              minRows: is_child ? 0 : 10,
+              minSpareRows: 0,
+              height: is_child ? '50vh' : '75vh',
+              // TODO: Workaround, possibly due to too small section column on child tables
+              // colWidths: is_child ? 256 : undefined, 
+              colWidths: 200,
+            },
+          });
+          //Object.assign(dh, handsOnDH);
+          handsOnDH.validator.useTargetClass(class_name);
+          handsOnDH['template'] = schema.classes[class_name];
+
+          // Set up the data structure (.template,.slots,.slot_names etc 
+          // based on LinkML class
+          this.useTemplate(handsOnDH, class_name); 
+
+          data_harmonizers[class_name] = handsOnDH;
+          handsOnDH.createHot();
+
+/*        FUTURE: separate dh from all .context schema/template data/functions
+
+          const dh = {
+            template: schema.classes[class_name]
+          }
+
+          // Set up the data structure (.template,.slots,.slot_names etc 
+          // based on LinkML class
+          this.useTemplate(dh, class_name); 
+
+          // Each selected DataHarmonizer is rendered here. 
+          // NOTE: this may be running twice?.
+          // in 1-M, different DataHarmonizers have fewer rows to start with
+          // Child tables are displayed differently so override the default
+          // HoT settings.
+          const dhSubroot = createDataHarmonizerContainer(dhId, index === 0);
+          const handsOnDH = new DataHarmonizer(dhSubroot, dh, {
+            loadingScreenRoot: document.body,
+            template_name: class_name,
+            schema: schema, // assign during constructor so validator can init on it.
+            hot_override_settings: {
+              minRows: is_child ? 0 : 10,
+              minSpareRows: 0,
+              height: is_child ? '50vh' : '75vh',
+              // TODO: Workaround, possibly due to too small section column on child tables
+              // colWidths: is_child ? 256 : undefined, 
+              colWidths: 200,
+            },
+          });
+
+          //Object.assign(dh, handsOnDH);
+          handsOnDH['template'] = schema.classes[class_name];
+          handsOnDH.validator.useTargetClass(class_name);
+          data_harmonizers[class_name] = handsOnDH;
+          handsOnDH.createHot();
+*/
+
+          if (is_child) {
+            /* Initialization of each child table is to hide all rows until
+             * parent primary key record is selected as a foreign key index.
+             * Note: user cut and paste of records doesn't change content of
+             * read-only fields, thankfully.
+            **/
+            handsOnDH.filterAll(handsOnDH); // Hides rows.
+            this.crudMakeForeignKeysReadOnly(handsOnDH, class_name);
+          }
+          // Top level class, so make it active.  There should only be one
+          else { 
+            this.setCurrentDataHarmonizer(class_name);
+          }
+
+        }
+      });
+    return data_harmonizers;
+  }
+
+/**
+   * Revise user interface elements to match template content
+   *
+   * @param {String} template_name: name of template within current schema
+   */
+  useTemplate(dh, template_name) {
+    dh.template_name = template_name;
+    dh.sections = []; // This will hold template's new data including table sections.
+    // let self = this;
+
+    const sectionIndex = new Map();
+
+    // Gets LinkML SchemaView() of given template
+    // problem: multiclass schemas means this template doesn't reflect the underlying structure anymore.
+    // const specification = this.schema.classes[template_name];
+    // Class contains inferred attributes ... for now pop back into slots
+
+    let attributes = Object.entries(
+      this.template.default.schema.classes
+    )
+      .filter(
+        ([cls_key]) =>
+          cls_key === dh.template_name
+      )
+      .reduce((acc, [, spec]) => {
+        return {
+          ...acc,
+          ...spec.attributes,
+        };
+      }, {});
+
+    /* Lookup each column in terms table. A term looks like:
+      is_a: "core field", 
+      title: "history/fire", 
+      slot_uri: "MIXS:0001086"
+      comments: (3) ['Expected value: date', 'Occurrence: 1', 'This field is used uniquely in: soil']
+      description: "Historical and/or physical evidence of fire"
+      examples: [{…}], 
+      multivalued: false, 
+      range: "date",
+      ...
+      */
+
+    for (let name in attributes) {
+      // ISSUE: a template's slot definition via SchemaView() currently
+      // doesn't get any_of or exact_mapping constructs. So we start
+      // with slot, then add class's slots reference.
+      let field = deepMerge(
+        attributes[name],
+        this.template.current.schema.slots[name]
+      );
+
+      //let field = attributes[name];
+      let section_title = null;
+
+      if ('slot_group' in field) {
+        // We have a field positioned within a section (or hierarchy)
+        section_title = field.slot_group;
+      } else {
+        if ('is_a' in field) {
+          section_title = field.is_a;
+        } else {
+          section_title = 'Generic';
+          console.warn("Template field doesn't have section: ", name);
+        }
+      }
+
+      // We have a field positioned within a section (or hierarchy)
+
+      if (!sectionIndex.has(section_title)) {
+        sectionIndex.set(section_title, sectionIndex.size);
+        let section_parts = {
+          title: section_title,
+          children: [],
+        };
+        const section =
+          this.template.current.schema['slots'][section_title];
+        if (section) {
+          Object.assign(section_parts, section);
+        }
+
+        dh.sections.push(section_parts);
+      }
+
+      let section = dh.sections[sectionIndex.get(section_title)];
+      let new_field = Object.assign(structuredClone(field), { section_title }); // shallow copy
+
+      // Some specs don't add plain english title, so fill that with name
+      // for display.
+      if (!('title' in new_field)) {
+        new_field.title = new_field.name;
+      }
+
+      // Default field type xsd:token allows all strings that don't have
+      // newlines or tabs
+      new_field.datatype = null;
+
+      let range_array = [];
+
+      // Multiple ranges allowed.  For now just accepting enumerations
+      if ('any_of' in new_field) {
+        for (let item of new_field.any_of) {
+          if (
+            item.range in this.template.default.schema.enums ||
+            item.range in this.template.default.schema.types ||
+            pascalToLowerWithSpaces(item.range) in
+              this.template.default.schema.enums ||
+            pascalToLowerWithSpaces(item.range) in
+              this.template.default.schema.types
+            // || Object.keys(this.template.current.schema.enums).some(k => k.indexOf('geo_loc_name') !== '-1') ||
+            // Object.keys(this.template.current.schema.types).some(k => k.indexOf('geo_loc_name') !== '-1')
+          ) {
+            range_array.push(item.range);
+          }
+        }
+      } else {
+        range_array.push(new_field.range);
+      }
+
+      // Parse slot's range(s)
+      for (let range of range_array) {
+        if (range === undefined) {
+          console.warn('field has no range', new_field.title);
+        }
+
+        // Range is a datatype?
+        const types = this.template.default.schema.types;
+        if (range in types || pascalToLowerWithSpaces(range) in types) {
+          const range_obj =
+            typeof types[range] !== 'undefined'
+              ? types[range]
+              : types[pascalToLowerWithSpaces(range)];
+
+          /* LinkML typically translates "string" to "uri":"xsd:string" 
+            // but this is problematic because that allows newlines which
+            // break spreadsheet saving of items in tsv/csv format. Use
+            // xsd:token to block newlines and tabs, and clean out leading 
+            // and trailing space. xsd:normalizedString allows lead and trai
+            // FUTURE: figure out how to accomodate newlines?
+            */
+          switch (range) {
+            case 'WhitespaceMinimizedString':
+              new_field.datatype = 'xsd:token';
+              break;
+            case 'string':
+              new_field.datatype = 'string';
+              break;
+            case 'Provenance':
+              new_field.datatype = 'Provenance';
+              break;
+            default:
+              new_field.datatype = range_obj.uri;
+            // e.g. 'time' and 'datetime' -> xsd:dateTime'; 'date' -> xsd:date
+          }
+        } else {
+          // If range is an enumeration ...
+          const enums = this.template.current.schema.enums;
+          if (range in enums || pascalToLowerWithSpaces(range) in enums) {
+            const range_obj =
+              typeof enums[range] !== 'undefined'
+                ? enums[range]
+                : enums[pascalToLowerWithSpaces(range)];
+
+            if (!('sources' in new_field)) {
+              new_field.sources = [];
+            }
+            if (!('flatVocabulary' in new_field)) {
+              new_field.flatVocabulary = [];
+            }
+            if (!('flatVocabularyLCase' in new_field)) {
+              new_field.flatVocabularyLCase = [];
+            }
+            if (!('permissible_values' in new_field)) {
+              new_field.permissible_values = {};
+            }
+
+            new_field.permissible_values[range] = range_obj.permissible_values;
+
+            new_field.sources.push(range);
+            // This calculates for each categorical field in schema.yaml a
+            // flat list of allowed values (indented to represent hierarchy)
+            let flatVocab = this.stringifyNestedVocabulary(
+              range_obj.permissible_values
+            );
+            new_field.flatVocabulary.push(...flatVocab);
+            // Lowercase version used for easy lookup/validation
+            new_field.flatVocabularyLCase.push(
+              ...flatVocab.map((val) => val.trim().toLowerCase())
+            );
+          } else {
+            // If range is a class ...
+            // multiple => 1-many complex object
+            if (range in this.template.current.schema.classes) {
+              if (range == 'quantity value') {
+                /* LinkML model for quantity value, along lines of https://schema.org/QuantitativeValue, e.g. xsd:decimal + unit
+                  PROBLEM: There are a variety of quantity values specified, some allowing units
+                  which would need to go in a second column unless validated as text within column.
+                  
+                  description: >-
+                  A simple quantity, e.g. 2cm
+                  attributes:
+                  verbatim:
+                  description: >-
+                  Unnormalized atomic string representation, should in syntax {number} {unit}
+                  has unit:
+                  description: >-
+                  The unit of the quantity
+                  slot_uri: qudt:unit
+                  has numeric value:
+                  description: >-
+                  The number part of the quantity
+                  range:
+                  double
+                  class_uri: qudt:QuantityValue
+                  mappings:
+                  - schema:QuantityValue
+                  */
+              }
+            }
+          }
+        } // End range parsing
+      }
+
+      // Provide default datatype if no other selected
+      if (!new_field.datatype) {
+        new_field.datatype = 'xsd:token';
+      }
+
+      // field.todos is used to store some date tests that haven't been
+      // implemented as rules yet.
+      if (new_field.datatype == 'xsd:date' && new_field.todos) {
+        // Have to feed any min/max date comparison back into min max value fields
+        for (const test of new_field.todos) {
+          if (test.substr(0, 2) == '>=') {
+            new_field.minimum_value = test.substr(2);
+          }
+          if (test.substr(0, 2) == '<=') {
+            new_field.maximum_value = test.substr(2);
+          }
+        }
+      }
+
+      /* Older DH enables mappings of one template field to one or more 
+        export format fields
+        */
+      this.setExportField(new_field, true);
+
+      /* https://linkml.io/linkml-model/docs/structured_pattern/
+       https://github.com/linkml/linkml/issues/674
+       Look up its parts in "settings", and assemble a regular 
+       expression for them to compile into "pattern" field. 
+       This augments basic datatype validation
+          structured_pattern:
+              syntax: "{float} {unit.length}"
+              interpolated: true   ## all {...}s are replaced using settings
+              partial_match: false
+      */
+      if ('structured_pattern' in new_field) {
+        switch (new_field.structured_pattern.syntax) {
+          case '{UPPER_CASE}':
+          case '{lower_case}':
+          case '{Title_Case}':
+            new_field.capitalize = true;
+        }
+        // TO DO: Do conversion here into pattern field.
+      }
+
+      // pattern is supposed to be exlusive to string_serialization
+      if ('pattern' in field && field.pattern.length) {
+        // Trap invalid regex
+        // Issue with NMDC MIxS "current land use" field pattern: "[ ....(all sorts of things) ]" syntax.
+        try {
+          new_field.pattern = new RegExp(field.pattern);
+        } catch (err) {
+          console.warn(
+            `TEMPLATE ERROR: Check the regular expression syntax for "${new_field.title}".`
+          );
+          console.error(err);
+          // Allow anything until regex fixed.
+          new_field.pattern = new RegExp(/.*/);
+        }
+      }
+      if (this.field_settings[name]) {
+        Object.assign(new_field, this.field_settings[name]);
+      }
+
+      if (field.annotations) {
+        new_field.annotations = field.annotations;
+      }
+
+      section.children.push(new_field);
+    } // End slot processing loop
+
+    // Sections and their children are sorted by .rank parameter if available
+    dh.sections.sort((a, b) => a.rank - b.rank);
+
+    // Sort kids in each section
+    for (let ptr in this.sections) {
+      dh.sections[ptr]['children'].sort((a, b) => a.rank - b.rank);
+    }
+
+    // Easy lookup arrays.
+    dh.slots = this.getSlots(dh);
+    dh.slot_names = dh.slots.map((slot) => slot.name);
+
+  }
+
+  /**
+   * Get a flat array of all fields (slot definitions) in TEMPLATE.
+   * @return {Array<Object>} Array of all objects under `children` in TEMPLATE.
+   */
+  getSlots(dh) {
+    let slots = Array.prototype.concat.apply(
+      [],
+      dh.sections.map((section) => section.children)
+    );
+    return slots;
+  }
+
+  /**
+   * Recursively flatten vocabulary into an array of strings, with each
+   * string's level of depth in the vocabulary being indicated by leading
+   * spaces.
+   * FUTURE possible functionality:
+   * Both validation and display of picklist item becomes conditional on
+   * other picklist item if "depends on" indicated in picklist item/branch.
+   * @param {Object} vocabulary See `vocabulary` fields in SCHEMA.
+   * @return {Array<String>} Flattened vocabulary.
+   */
+  stringifyNestedVocabulary(vocab_list) {
+    let ret = [];
+    let stack = [];
+    for (const pointer in vocab_list) {
+      let choice = vocab_list[pointer];
+      let level = 0;
+      if ('is_a' in choice) {
+        level = stack.indexOf(choice.is_a) + 1;
+        stack.splice(level + 1, 1000, choice.text);
+      } else {
+        stack = [choice.text];
+      }
+
+      this.setExportField(choice, false);
+
+      ret.push('  '.repeat(level) + choice.text);
+    }
+    return ret;
+  }
+
+  setExportField(field, as_field) {
+    if (field.exact_mappings) {
+      field.exportField = {};
+      for (let item of field.exact_mappings) {
+        let ptr = item.indexOf(':');
+        if (ptr != -1) {
+          const prefix = item.substr(0, ptr);
+          if (!(prefix in field.exportField)) {
+            field.exportField[prefix] = [];
+          }
+
+          const mappings = item.substr(ptr + 1);
+          for (let mapping of mappings.split(';')) {
+            mapping = mapping.trim();
+            const conversion = {};
+            //A colon alone means to map value to empty string
+            if (mapping == ':') {
+              conversion.value = '';
+            }
+            //colon with contents = field & value
+            else {
+              mapping = decodeURIComponent(mapping);
+              if (mapping.indexOf(':') != -1) {
+                const binding = mapping.split(':');
+                binding[0] = binding[0].trim();
+                binding[1] = binding[1].trim();
+                if (binding[0] > '') {
+                  conversion.field = binding[0];
+                }
+                if (binding[1] > '') {
+                  conversion.value = binding[1];
+                }
+              }
+              //No colon means its just field or value
+              else if (as_field == true) {
+                conversion.field = mapping;
+              } else {
+                conversion.value = mapping;
+              }
+            }
+            field.exportField[prefix].push(conversion);
+          }
+        }
+      }
+    }
   }
 
   /**
@@ -241,149 +777,28 @@ export default class AppContext {
     });
   }
 
-  /**
-   * Creates Data Harmonizers from the schema tree.
-   * FUTURE: Make this a 2 pass thing.  First pass creates every dataharmonizer
-   * data structure; 2nd pass initializes handsontable part.
-   * @param {Object} schema_tree - The schema tree from which to create Data Harmonizers.
-   * @returns {Object} An object mapping the class names to their respective Data Harmonizer objects.
-   */
-  makeDHsFromRelations(schema, template_name) {
-    let data_harmonizers = {};
-    Object.entries(schema.classes)
-      // Container class is only used in input and output file coordination.
-      .filter(([class_name]) => class_name == template_name || this.relations[template_name].dependents.has(class_name))
-      .forEach((obj, index) => {
-        if (obj.length > 0) {
 
-          const [class_name, tree_obj] = obj;
-
-          // if it shares a key with another class which is its parent, this DH must be a child
-          const is_child = this.crudGetParents(class_name);
-
-          // Doesn't build if tab id already exists.
-          const dhId = `data-harmonizer-grid-${index}`;
-          if (!document.getElementById(dhId)) {
-            const dhSubroot = createDataHarmonizerContainer(dhId, index === 0);
-
-            const dhTab = createDataHarmonizerTab(dhId, class_name, index === 0);
-            dhTab.addEventListener('click', () => {
-              $(document).trigger('dhTabChange', {
-                specName: class_name,
-              });
-            });
-
-            // Each selected DataHarmonizer is rendered here. 
-            // NOTE: this may be running twice?.
-            // in 1-M, different DataHarmonizers have fewer rows to start with
-            // Child tables are displayed differently so override the default
-            // HoT settings.
-            const dh = new DataHarmonizer(dhSubroot, this, {
-              loadingScreenRoot: document.body,
-              template_name: class_name,
-              schema: schema, // assign during constructor so validator can init on it.
-              hot_override_settings: {
-                minRows: is_child ? 0 : 10,
-                minSpareRows: 0,
-                height: is_child ? '50vh' : '75vh',
-                // TODO: Workaround, possibly due to too small section column on child tables
-                // colWidths: is_child ? 256 : undefined, 
-                colWidths: 200,
-              },
-            });
-
-            dh.template = schema.classes[class_name];
-            // Set up the data structure (.template,.slots,.slot_names etc 
-            // based on LinkML class
-            dh.useTemplate(class_name); 
-            dh.createHot();
-            dh.validator.useTargetClass(class_name);
-            
-            data_harmonizers[class_name] = dh;
-
-            if (is_child) {
-              /* Initialization of each child table is to hide all rows until
-               * parent primary key record is selected as a foreign key index.
-               * Note: user cut and paste of records doesn't change content of
-               * read-only fields, thankfully.
-              **/
-              dh.filterAll(); // Hides rows.
-              this.crudMakeForeignKeysReadOnly(dh, class_name);
-            }
-            // Top level class, so make it active.  There should only be one
-            else { 
-              this.setCurrentDataHarmonizer(class_name);
-            }
-          }
-        }
-      });
-    delete data_harmonizers[undefined]; // How does this occur?
-    return data_harmonizers;
-  }
-
-  /*
-   * Return initialized, rendered dataHarmonizer instances rendered in order
-   * of their appearance as classes in schema. If a template name (class name)
-   * is provided, only include that one and any dependent tables in its schema.
-   * Each class has one or more slots (attributes) which each contain a 
-   * foreign_key which determines parent-child relationships.
-   *
-   * Uses this.relations to construct list of main DH class and its dependant 
-   * classes.
-   *
-   * Processing each class into DataHarmonizer tables:
-   * - One DH instance per class: this.dhs[template_name]
-   *   - In rendered DH view, filter its table rows by the foreign key slots 
-   *     featured in that class.
-   * - Navigation: One tab per DH instance.
-   */
-  async setupDataHarmonizers({
-    data_harmonizers = {},
-    locale = null,
-    forced_schema = null
-  }) {
-    
-    this.clearInterface();
-
-    const [schema_path, template_name] = this.appConfig.template_path.split('/');
-
-    this.template = await Template.create(schema_path, {forced_schema});
-
-    // Empty case will occur when template_path doesn't correspond to a built template
-    this.export_formats = forced_schema ? {} : await getExportFormats(schema_path);
-
-    if (locale !== null) {
-      this.template.updateLocale(locale);
-    }
-
-    const schema =
-      locale !== null
-        ? this.template.localized.schema
-        : this.template.default.schema;
-
-    // FUTURE slot_ptr version of self.context.relations for efficiency?
-    this.relations = this.crudGetRelations(schema);
-    console.log("GETRELATIONS",this.relations)
-    this.schema = schema;
-
-    // Merges any existing dataharmonizer instances with the ones newly created.
-    this.dhs = {
-      ...data_harmonizers,
-      ...this.makeDHsFromRelations(schema, template_name),
-    };
-
-
-    /*
-    for (const dh in this.dhs)
-        this.relations[template_name].dependents = this.crudGetDependents(dh.template_name)
-    */
-    return this;
-
-  }
 
   /**************************** CRUD RELATED METHODS *************************/
+  /** 
+   * General behaviour
+   * 
+   * WRT rows with keys pointing to foreign keys, a user:
+   * - Can't input directly into foreign key cells.  
+   * - Can delete a row with foreign key(s)
+   * - Can add a new row only if foreign key(s) table(s) are
+   *   focused on key fields that have values.
+   *
+   * WRT rows which are foreign keys to other tables, a user:
+   * - Can update or delete a key field which then cascades to
+   *   other child table(s). This might trigger violation of unique_keys!
+   * 
+   * Note: user cut and paste of records doesn't change read-only
+   * fields.
+  */ 
 
-    /* Establish an easy bidirectional lookup between schema classes that have
+  /**
+   * Establish an easy bidirectional lookup between schema classes that have
    * foreign keys. Usually one table will have 1 or n slots keyed to one other
    * table's slots.  Sometimes a table may have two slots keyed to respective
    * other tables.  
@@ -436,99 +851,61 @@ export default class AppContext {
     }
 
   */
-
-  /* WRT rows with keys pointing to foreign keys, a user:
-   * - Can't input directly into foreign key cells.  
-   * - Can delete a row with foreign key(s)
-   * - Can add a new row only if foreign key(s) table(s) are
-   *   focused on key fields that have values.
-   *
-   * WRT rows which are foreign keys to other tables, a user:
-   * - Can update or delete a key field which then cascades to
-   *   other child table(s). This might trigger violation of unique_keys!
-   * 
-   * Note: user cut and paste of records doesn't change read-only
-   * fields.
-  */ 
-
   crudGetRelations(schema) {
     let relations = {};
     Object.entries(schema.classes).forEach(([class_name, class_obj]) => {
-      relations[class_name] = {};
+      if (class_name != 'Container' && class_name != 'dh_interface') {
+        relations[class_name] = {};
 
-      Object.entries(class_obj.attributes ?? {}).forEach(([slot_name, attribute]) => {
-        if (attribute.annotations?.foreign_key?.value) {
-          let foreign_class, foreign_slot_name;
-          let key = attribute.annotations.foreign_key.value;
-          // FIRST, TRY GETTING PARENT CLASS VIA foreign class with dot prefix:
-          if (key.includes('.'))
-            [foreign_class, foreign_slot_name] = key.split('.',2);
-          // ALTERNATELY (UNTESTED) USE range of slot having reference to a class:
-          // NOTE: this doesn't handle multiple classes in range.
-          else if (attribute.range in schema.classes) {
-            foreign_slot_name = key;
-            foreign_class = attribute.range;
+        Object.entries(class_obj.attributes ?? {}).forEach(([slot_name, attribute]) => {
+          if (attribute.annotations?.foreign_key?.value) {
+            let foreign_class, foreign_slot_name;
+            let key = attribute.annotations.foreign_key.value;
+            // FIRST, TRY GETTING PARENT CLASS VIA foreign class with dot prefix:
+            if (key.includes('.'))
+              [foreign_class, foreign_slot_name] = key.split('.',2);
+            // ALTERNATELY (UNTESTED) USE range of slot having reference to a class:
+            // NOTE: this doesn't handle multiple classes in range.
+            else if (attribute.range in schema.classes) {
+              foreign_slot_name = key;
+              foreign_class = attribute.range;
+            }
+            else {
+              console.log("Class", class_name, "has slot", slot_name, "foreign key", attribute.annotations?.foreign_key?.value, "but no target class information in key or slot range.");
+              return;
+            }
+            Object.assign(relations[class_name],  
+              {parent: {[foreign_class]: {[slot_name]: foreign_slot_name}}});
+            // And reverse relation
+            Object.assign(relations[foreign_class], 
+              {child: {[class_name]: {[foreign_slot_name]: slot_name}}});
+            // And dependent slots via foreign key relations.
+            Object.assign(relations[class_name], 
+              {dependent_slots: {[slot_name]: {parent: foreign_class, slot: foreign_slot_name}}});
           }
-          else {
-            console.log("Class", class_name, "has slot", slot_name, "foreign key", attribute.annotations?.foreign_key?.value, "but no target class information in key or slot range.");
-            return;
-          }
-          Object.assign(relations[class_name],  
-            {parent: {[foreign_class]: {[slot_name]: foreign_slot_name}}});
-          // And reverse relation
-          Object.assign(relations[foreign_class], 
-            {child: {[class_name]: {[foreign_slot_name]: slot_name}}});
-          // And dependent slots via foreign key relations.
-          Object.assign(relations[class_name], 
-            {dependent_slots: {[slot_name]: {parent: foreign_class, slot: foreign_slot_name}}});
-        }
-      });
-
-      // Now do unique keys in class_obj which might not be mentioned in 
-      // foreign_class relationships.
-      Object.entries(class_obj.unique_keys ?? {}).forEach(([key_name, key_obj]) => {
-        Object.entries(key_obj.unique_key_slots ?? {}).forEach(([index, slot_name]) => {
-          Object.assign(relations[class_name], 
-            {unique_key_slots: {[slot_name]: {[key_name]: true}}});
         });
-      });
+
+        // Now do unique keys in class_obj which might not be mentioned in 
+        // foreign_class relationships.
+        Object.entries(class_obj.unique_keys ?? {}).forEach(([key_name, key_obj]) => {
+          Object.entries(key_obj.unique_key_slots ?? {}).forEach(([index, slot_name]) => {
+            Object.assign(relations[class_name], 
+              {unique_key_slots: {[slot_name]: {[key_name]: true}}});
+          });
+        });
+      }
     });
 
-    // 2nd pass, since all dependents have been created, AND this must produce
-    // a carefully organized MAP {[class_name]: dh...} which functions can 
-    // process in left-to-right order to account for primary/foreign key 
-    // relations.
-    for (const class_name in schema.classes) {
-      relations[class_name].dependents = this.crudGetDependents(class_name);
-    }
     return relations;
   }
 
   crudGetParents(template_name) {
-    return this.relations?.[template_name]?.parent;
+    return this.relations[template_name]?.parent;
   }
 
   crudGetChildren(template_name) {
-    return this.relations?.[template_name]?.child;
+    return this.relations[template_name]?.child;
   }
-
-  /*
-  // Determine if parent relationships link class_name to given ancestor
-  crudIsAncestor(class_names, ancestor_name) {
-    let subject_name = class_names.shift();
-    if (!subject_name)
-      return false;
-    if (subject_name == ancestor_name)
-      return true;
-    if (this.relations[subject_name]?.parent) {
-      if (this.relations[subject_name].parent[ancestor_name])
-        return true;
-      // Ancestor wasn't one of parents, so try search on all ancestors.
-      class_names.push(Object.values(this.relations[subject_name].parent));
-    }
-    return this.crudIsAncestor(class_names, ancestor_name)
-  }
-  */
 
   /** Retrieves ORDERED DICT of tables that ultimately have given template as
    * a foreign key, IN ORDER.  Whether it is to enact cascading visibility, update or
@@ -542,12 +919,13 @@ export default class AppContext {
    * @return {Object} class_names where each descendent class_name is mentioned.
    */
   crudGetDependents(class_name, ptr=-1, stack = new Map()) {
-    if (ptr == stack.length) 
+    if (ptr == stack.length) {
       return stack;
+    }
     const children = this.crudGetChildren(class_name);
-    if (!children)
+    if (!children) {// catches "undefined"
       return (stack)
-
+    }
     // Add each child to end of stack.
     let next_name = '';
     for (const dependent_name in children) {
@@ -558,17 +936,12 @@ export default class AppContext {
     return this.crudGetDependents(next_name, ptr+1, stack);
   }
 
-  crudCheckDependency (self, row, dependent_name, changes) {
-    let keyVals = {};
-    let change_log = '';
-    let found = false;
-    return [found, keyVals, change_log];
-  }
-
   /* For given class, refresh view of all dependent tables that have a direct
    * or indirect foreign key relationship to given class.
    * Performance might show up as an issue later if lots of long dependent
    * tables to re-render.
+   * FUTURE: show enabled/disabled dependent tab based on whether parent
+   * table has focused primary key? 
    */
   crudFilterDependentViews(class_name) {
     for (var [dependent_name, obj] of this.relations[class_name].dependents) { 
@@ -603,11 +976,11 @@ export default class AppContext {
     // parent(s) keys?
     if (errors.length == 0) {
       rowsToHide = rowsToHide.filter((row) => {
-        for (let [slot_name, value] of Object.entries(required_selections)) {
+        for (const [slot_name, value] of Object.entries(required_selections)) {
           const col = this.dhs[class_name].getColumnIndexByFieldName(slot_name);
           const cellValue = hotInstance.getDataAtCell(row, col);
           // Null value test in case where parent or subordinate field is null.
-          if (cellValue != value)
+          if (cellValue != value.value)
             return true;
         }
       });
@@ -640,10 +1013,10 @@ export default class AppContext {
       Object.entries(parent).forEach(([slot_name, foreign_slot_name]) => {
         const value = this.crudGetSelectedTableSlotValue(parent_name, foreign_slot_name);
         if (value == null) {
-          errors += `<li> <b>${parent_name}</b> (${foreign_slot_name})</li>`;
+          errors += `\n- ${parent_name}.${foreign_slot_name}: _____`;
         }
         else {
-          required_selections[slot_name] = value;
+          required_selections[slot_name] = {value:value, parent:parent_name, slot:foreign_slot_name};
         }
       });
     });
@@ -671,6 +1044,57 @@ export default class AppContext {
 
   };
 
+  /** TEST FOR CHANGE TO DEPENDENT TABLE FOREIGN KEY(S)
+   * CASE 1: A shift from a complete foreign key to:
+   * A) Another complete combination.
+   *    1) duplicate?  
+   *       a) MERGE?
+   *    2) unique:
+   *       Prompt to change all underlying records?
+   * B) Deleted part
+   *    BLOCK.  Suggest deleting row.
+   * C) Cleared out.
+   *    Block - same as B).
+
+   * - changed key (part) -> prompt user to use "right click: Change primary key option"
+   *
+   * CASE 2: Shift from incomplete foreign key to complete one:
+   * - This doesn't necessitate any action except say for a refresh
+   *   of dependent table in case where we allow dependent to have 
+   *   "loose cannon" foreign key references.
+   * ISSUE IS THIS WOULD HAVE TO CASCADE KEY SENSITIVITY.
+   * WHERE DOES KEY SENSITIVITY RESPONSIBILITY END?
+   * BEST TO RECURSIVELY CHECK DEPENDENCIES
+   * @param {String} class_name name of root template/class to check
+   * @param {Integer} row of parent template to check for primary key values.
+   * @param {Array} changes to primary key values on that row????
+   * @param {Boolean} update indicating whether to make changes to data.
+   */
+  crudGetDependentUpdates(class_name, row, update=false) {
+    let change_log = '';
+    for (const [dependent_name, obj] of this.relations[class_name].dependents) {
+      const parents = this.crudGetParents(dependent_name);
+      // Checks to see which selection area of each parent is active (usually
+      // just one parent but sometimes a table is dependent on two others)
+      // GETS RECORDS THAT MATCH EXISTING KEYS, for which new KEY(S) pertain.
+      let [required_selections, errors] = this.crudGetForeignKeyValues(parents);
+      // Ignore tables with errors as these indicate dependent tables with 
+      // incomplete foreign keys.
+      // Note: user can "rescue" visibility of dependent records that already
+      // exist with final key value(s).
+      if (!errors)
+        for (const [slot_name, value] of Object.entries(required_selections)) {
+          change_log = change_log + `\n - ${dependent_name} record(s) with ${slot_name}: ${value.value}`;
+        }
+    }
+
+    // Here we have complete unique_keys search to find relate records by.
+    if (change_log > '') {
+        change_log = `Your key change on ${class_name} row ${parseInt(row)+1} would also change existing dependent table records! Do you want to continue?  Check:\n ${change_log}`;
+    }
+    return change_log;
+  }
+
   /* If dependent table is connected to key in parent table which is given by
    * selected row (i.e. dependent foreign keys = values on this row) AND that
    * key is unique in this table, then deletion of this row should lead to 
@@ -682,9 +1106,8 @@ export default class AppContext {
    */ 
   crudGetDependentDeletes(class_name, do_delete = false) {
     let deletes = [];
-    let warning_text = 'WARNING: If you proceed, this will include deletion of one\n or more dependent records:\n\n';
+    let warning_text = 'WARNING: If you proceed, this will include deletion of one\n or more dependent records:\n';
 
-    const dependents = this.crudGetDependents(class_name);
     /* 1) Determine if each dependent's foreign keys are satisfied - each must
      *    have a value matching class_name slot's value; ALSO, This set of 
      *    values must be unique to a single row if in parent table. If it 
@@ -699,7 +1122,7 @@ export default class AppContext {
      *    parent.
      */
     let found_dependent_rows = false;
-    for (const dependent_name in dependents) {
+    for (const [dependent_name, obj] of this.relations[class_name].dependents) {
 
       const ddh = this.dhs[dependent_name];
       const parents = this.crudGetParents(dependent_name);
@@ -735,7 +1158,7 @@ export default class AppContext {
           else if (found_rows.length > 0) {
             found_dependent_rows = true;
             let key_vals = Object.entries(d_KeyVals).join("\n").replace(',',': ');
-            warning_text += `- ${found_rows.length} ${dependent_name} records with key:\n      ${key_vals}`;
+            warning_text += `\n - ${found_rows.length} ${dependent_name} records with key:\n      ${key_vals}`;
           }
         }
       });
