+++ lib/DataHarmonizer.js
@@ -49,11 +49,9 @@ import unmappedHeadersModal from './unmappedHeadersModal.html';
 import fieldDescriptionsModal from './fieldDescriptionsModal.html';
 
 import HelpSidebar from './HelpSidebar';
-
-// NOTE: this is odd! package.json is a developer file. why should a UI component care about it?
-import pkg from '../package.json';
 import Validator from './Validator';
 
+import pkg from '../package.json'; // Retrieves package version id.
 const VERSION = pkg.version;
 const VERSION_TEXT = 'DataHarmonizer v' + VERSION;
 
@@ -86,21 +84,25 @@ Handsontable.cellTypes.registerCellType('dh.time', {
   renderer: Handsontable.renderers.getRenderer('autocomplete'),
 });
 
+/**
+ * An instance of DataHarmonizer has a schema, a domElement, and a 
+ * handsontable .hot object
+ */
 class DataHarmonizer {
-  //An instance of DataHarmonizer has a schema, a domElement, and a handsontable .hot object
+
   root = null;
   template_name = null;
   schema = null; // Schema holding all templates
   sections = null; // Specific template from schema
+  slots = null;
+  slot_names = null;
   hot = null;
   menu = null;
   export_formats = {}; // Formats that a given template can export to.
   invalid_cells = null;
   // Currently selected cell range[row,col,row2,col2]
   current_selection = [null, null, null, null];
-  field_settings = {};
-  fields = [];
-  template_unique_keys = [];
+
 
   constructor(root, context, options = {}) {
     // table schema
@@ -189,335 +191,6 @@ class DataHarmonizer {
 
 
   /**
-   * Revise user interface elements to match template content
-   *
-   * @param {String} template_name: name of template within current schema
-   */
-  useTemplate(template_name) {
-    this.template_name = template_name;
-    this.sections = []; // This will hold template's new data including table sections.
-    // let self = this;
-
-    const sectionIndex = new Map();
-
-    // Gets LinkML SchemaView() of given template
-    // problem: multiclass schemas means this template doesn't reflect the underlying structure anymore.
-    // const specification = this.schema.classes[template_name];
-    // Class contains inferred attributes ... for now pop back into slots
-
-    let attributes = Object.entries(
-      this.context.template.default.schema.classes
-    )
-      .filter(
-        ([cls_key]) =>
-          cls_key === this.template_name
-      )
-      .reduce((acc, [, spec]) => {
-        return {
-          ...acc,
-          ...spec.attributes,
-        };
-      }, {});
-
-    /* Lookup each column in terms table. A term looks like:
-      is_a: "core field", 
-      title: "history/fire", 
-      slot_uri: "MIXS:0001086"
-      comments: (3) ['Expected value: date', 'Occurrence: 1', 'This field is used uniquely in: soil']
-      description: "Historical and/or physical evidence of fire"
-      examples: [{…}], 
-      multivalued: false, 
-      range: "date",
-      ...
-      */
-
-    for (let name in attributes) {
-      // ISSUE: a template's slot definition via SchemaView() currently
-      // doesn't get any_of or exact_mapping constructs. So we start
-      // with slot, then add class's slots reference.
-      let field = deepMerge(
-        attributes[name],
-        this.context.template.current.schema.slots[name]
-      );
-
-      //let field = attributes[name];
-      let section_title = null;
-
-      if ('slot_group' in field) {
-        // We have a field positioned within a section (or hierarchy)
-        section_title = field.slot_group;
-      } else {
-        if ('is_a' in field) {
-          section_title = field.is_a;
-        } else {
-          section_title = 'Generic';
-          console.warn("Template field doesn't have section: ", name);
-        }
-      }
-
-      // We have a field positioned within a section (or hierarchy)
-
-      if (!sectionIndex.has(section_title)) {
-        sectionIndex.set(section_title, sectionIndex.size);
-        let section_parts = {
-          title: section_title,
-          children: [],
-        };
-        const section =
-          this.context.template.current.schema['slots'][section_title];
-        if (section) {
-          Object.assign(section_parts, section);
-        }
-
-        this.sections.push(section_parts);
-      }
-
-      let section = this.sections[sectionIndex.get(section_title)];
-      let new_field = Object.assign(structuredClone(field), { section_title }); // shallow copy
-
-      // Some specs don't add plain english title, so fill that with name
-      // for display.
-      if (!('title' in new_field)) {
-        new_field.title = new_field.name;
-      }
-
-      // Default field type xsd:token allows all strings that don't have
-      // newlines or tabs
-      new_field.datatype = null;
-
-      let range_array = [];
-
-      // Multiple ranges allowed.  For now just accepting enumerations
-      if ('any_of' in new_field) {
-        for (let item of new_field.any_of) {
-          if (
-            item.range in this.context.template.default.schema.enums ||
-            item.range in this.context.template.default.schema.types ||
-            pascalToLowerWithSpaces(item.range) in
-              this.context.template.default.schema.enums ||
-            pascalToLowerWithSpaces(item.range) in
-              this.context.template.default.schema.types
-            // || Object.keys(this.context.template.current.schema.enums).some(k => k.indexOf('geo_loc_name') !== '-1') ||
-            // Object.keys(this.context.template.current.schema.types).some(k => k.indexOf('geo_loc_name') !== '-1')
-          ) {
-            range_array.push(item.range);
-          }
-        }
-      } else {
-        range_array.push(new_field.range);
-      }
-
-      // Parse slot's range(s)
-      for (let range of range_array) {
-        if (range === undefined) {
-          console.warn('field has no range', new_field.title);
-        }
-
-        // Range is a datatype?
-        const types = this.context.template.default.schema.types;
-        if (range in types || pascalToLowerWithSpaces(range) in types) {
-          const range_obj =
-            typeof types[range] !== 'undefined'
-              ? types[range]
-              : types[pascalToLowerWithSpaces(range)];
-
-          /* LinkML typically translates "string" to "uri":"xsd:string" 
-            // but this is problematic because that allows newlines which
-            // break spreadsheet saving of items in tsv/csv format. Use
-            // xsd:token to block newlines and tabs, and clean out leading 
-            // and trailing space. xsd:normalizedString allows lead and trai
-            // FUTURE: figure out how to accomodate newlines?
-            */
-          switch (range) {
-            case 'WhitespaceMinimizedString':
-              new_field.datatype = 'xsd:token';
-              break;
-            case 'string':
-              new_field.datatype = 'string';
-              break;
-            case 'Provenance':
-              new_field.datatype = 'Provenance';
-              break;
-            default:
-              new_field.datatype = range_obj.uri;
-            // e.g. 'time' and 'datetime' -> xsd:dateTime'; 'date' -> xsd:date
-          }
-        } else {
-          // If range is an enumeration ...
-          const enums = this.context.template.current.schema.enums;
-          if (range in enums || pascalToLowerWithSpaces(range) in enums) {
-            const range_obj =
-              typeof enums[range] !== 'undefined'
-                ? enums[range]
-                : enums[pascalToLowerWithSpaces(range)];
-
-            if (!('sources' in new_field)) {
-              new_field.sources = [];
-            }
-            if (!('flatVocabulary' in new_field)) {
-              new_field.flatVocabulary = [];
-            }
-            if (!('flatVocabularyLCase' in new_field)) {
-              new_field.flatVocabularyLCase = [];
-            }
-            if (!('permissible_values' in new_field)) {
-              new_field.permissible_values = {};
-            }
-
-            new_field.permissible_values[range] = range_obj.permissible_values;
-
-            new_field.sources.push(range);
-            // This calculates for each categorical field in schema.yaml a
-            // flat list of allowed values (indented to represent hierarchy)
-            let flatVocab = this.stringifyNestedVocabulary(
-              range_obj.permissible_values
-            );
-            new_field.flatVocabulary.push(...flatVocab);
-            // Lowercase version used for easy lookup/validation
-            new_field.flatVocabularyLCase.push(
-              ...flatVocab.map((val) => val.trim().toLowerCase())
-            );
-          } else {
-            // If range is a class ...
-            // multiple => 1-many complex object
-            if (range in this.context.template.current.schema.classes) {
-              if (range == 'quantity value') {
-                /* LinkML model for quantity value, along lines of https://schema.org/QuantitativeValue, e.g. xsd:decimal + unit
-                  PROBLEM: There are a variety of quantity values specified, some allowing units
-                  which would need to go in a second column unless validated as text within column.
-                  
-                  description: >-
-                  A simple quantity, e.g. 2cm
-                  attributes:
-                  verbatim:
-                  description: >-
-                  Unnormalized atomic string representation, should in syntax {number} {unit}
-                  has unit:
-                  description: >-
-                  The unit of the quantity
-                  slot_uri: qudt:unit
-                  has numeric value:
-                  description: >-
-                  The number part of the quantity
-                  range:
-                  double
-                  class_uri: qudt:QuantityValue
-                  mappings:
-                  - schema:QuantityValue
-                  */
-              }
-            }
-          }
-        } // End range parsing
-      }
-
-      // Provide default datatype if no other selected
-      if (!new_field.datatype) {
-        new_field.datatype = 'xsd:token';
-      }
-
-      // field.todos is used to store some date tests that haven't been
-      // implemented as rules yet.
-      if (new_field.datatype == 'xsd:date' && new_field.todos) {
-        // Have to feed any min/max date comparison back into min max value fields
-        for (const test of new_field.todos) {
-          if (test.substr(0, 2) == '>=') {
-            new_field.minimum_value = test.substr(2);
-          }
-          if (test.substr(0, 2) == '<=') {
-            new_field.maximum_value = test.substr(2);
-          }
-        }
-      }
-
-      /* Older DH enables mappings of one template field to one or more 
-        export format fields
-        */
-      this.setExportField(new_field, true);
-
-      /* https://linkml.io/linkml-model/docs/structured_pattern/
-       https://github.com/linkml/linkml/issues/674
-       Look up its parts in "settings", and assemble a regular 
-       expression for them to compile into "pattern" field. 
-       This augments basic datatype validation
-          structured_pattern:
-              syntax: "{float} {unit.length}"
-              interpolated: true   ## all {...}s are replaced using settings
-              partial_match: false
-      */
-      if ('structured_pattern' in new_field) {
-        switch (new_field.structured_pattern.syntax) {
-          case '{UPPER_CASE}':
-          case '{lower_case}':
-          case '{Title_Case}':
-            new_field.capitalize = true;
-        }
-        // TO DO: Do conversion here into pattern field.
-      }
-
-      // pattern is supposed to be exlusive to string_serialization
-      if ('pattern' in field && field.pattern.length) {
-        // Trap invalid regex
-        // Issue with NMDC MIxS "current land use" field pattern: "[ ....(all sorts of things) ]" syntax.
-        try {
-          new_field.pattern = new RegExp(field.pattern);
-        } catch (err) {
-          console.warn(
-            `TEMPLATE ERROR: Check the regular expression syntax for "${new_field.title}".`
-          );
-          console.error(err);
-          // Allow anything until regex fixed.
-          new_field.pattern = new RegExp(/.*/);
-        }
-      }
-      if (this.field_settings[name]) {
-        Object.assign(new_field, this.field_settings[name]);
-      }
-
-      if (field.annotations) {
-        new_field.annotations = field.annotations;
-      }
-
-      section.children.push(new_field);
-    } // End slot processing loop
-
-    // Sections and their children are sorted by .rank parameter if available
-    this.sections.sort((a, b) => a.rank - b.rank);
-
-    // Sort kids in each section
-    for (let ptr in this.sections) {
-      this.sections[ptr]['children'].sort((a, b) => a.rank - b.rank);
-    }
-
-    // Easy lookup arrays.
-    this.slots = this.getFields();
-    this.slot_names = this.slots.map((field) => field.name);
-
-  }
-
-
-  /**
-   * Get a flat array of all fields (slot definitions) in TEMPLATE.
-   * @return {Array<Object>} Array of all objects under `children` in TEMPLATE.
-   */
-  getFields() {
-    let fields = Array.prototype.concat.apply(
-      [],
-      this.sections.map((section) => section.children)
-    );
-
-    /* NOTE: Possibly a result of attempting to add other slots besides one in
-     * template.
-
-    fields =
-      this.slot_names.length > 0
-        ? fields.filter((field) => this.slot_names.includes(field.name))
-        : fields;
-    */
-    return fields;
-  }
-
-  /**
    * Open file specified by user.
    * Only opens `xlsx`, `xlsx`, `csv` and `tsv` files. Will launch the specify
    * headers modal if the file's headers do not match the grid's headers.
@@ -594,7 +267,7 @@ class DataHarmonizer {
   }
 
   newHotFile() {
-    this.runBehindLoadingScreen(
+    this.context.runBehindLoadingScreen(
       function () {
         this.createHot();
       }.bind(this)
@@ -618,8 +291,8 @@ class DataHarmonizer {
       licenseKey: 'non-commercial-and-evaluation',
     });
 
-    let fields = this.slots;
-    if (fields.length) {
+    if (this.slots) {
+      /*
       this.clipboardCache = '';
       this.clipboardCoordCache = {
         'CopyPaste.copy': {},
@@ -646,6 +319,7 @@ class DataHarmonizer {
           self.clipboardCoordCache['CopyPaste.paste'] = coords[0];
           self.clipboardCoordCache['Action:CopyPaste'] = 'paste';
         },
+
         contextMenu: [
           'copy',
           'cut',
@@ -656,25 +330,26 @@ class DataHarmonizer {
               return self.clipboardCache.length === 0;
             },
             callback: function () {
-              var plugin = this.getPlugin('copyPaste');
+              //var plugin = this.getPlugin('copyPaste');
 
-              this.listen();
+              //this.listen();
               // BUG: It seems like extra lf is added by sheetclip, causing
               // empty last row to be added and pasted. Exception is pasting
               // of single empty cell
-              if (self.clipboardCache.length > 0)
-                self.clipboardCache = self.clipboardCache.slice(0, -1);
-              plugin.paste(self.clipboardCache);
+              //if (self.clipboardCache.length > 0)
+              //  self.clipboardCache = self.clipboardCache.slice(0, -1);
+              //plugin.paste(self.clipboardCache);
             },
           },
           'remove_row',
           'row_above',
           'row_below',
         ],
-      };
 
+      };
+      */
       const hot_settings = {
-        ...hot_copy_paste_settings,
+        //...hot_copy_paste_settings,
         data: data, // Enables true reset
         nestedHeaders: this.getNestedHeaders(),
         autoColumnSize: true, // Enable automatic column size calculation
@@ -688,11 +363,12 @@ class DataHarmonizer {
             key: 'remove_row',
             name: 'Remove row',
             callback: function () {
-              //Enables removal of a row and all subordinate rows in other DH tables.
+              // Enables removal of a row and all dependent table rows.
               // If there are 1-many cascading deletes, verify if that's ok.
               let warning = self.context.crudGetDependentDeletes(self.template_name);
               if (!warning || confirm (warning)) {
-                if (warning)
+                // User has seen the warning and has confirmed ok to proceed.
+                if (warning) 
                   // "True" parameter triggers deletions
                   self.context.crudGetDependentDeletes(self.template_name, true);
                 let selection = self.hot.getSelected()[0][0];
@@ -703,7 +379,7 @@ class DataHarmonizer {
           {
             key: 'row_above',
             name: 'Insert row above',
-            callback: function () {
+            callback: function (action,selection,event) {
               // Ensuring that rows inserted into foreign-key dependent tables
               // take in the appropriate focused foreign-key values on creation.
               self.addRows('insert_row_above', 1, self.hot.getSelected()[0][0]);
@@ -776,7 +452,11 @@ class DataHarmonizer {
         invalidCellClassName: '',
         licenseKey: 'non-commercial-and-evaluation',
 
-        /* Before carrying out user-specificed changes to one or more cells,
+        //beforePaste:
+        //afterPaste:
+
+        /**
+         * Before carrying out user-specificed changes to one or more cells,
          * validate changes to prevent:
          * - any unique (key) column slot from getting a duplicate.
          * - any unique_key slot combination from getting a duplicate.
@@ -784,22 +464,35 @@ class DataHarmonizer {
          *   - includes catching unintended deletions or updates.
          * Also check to see if changes should trigger other rules.
          *
-         * One assumption: no functioning key slot can have null values.
+         * One assumption: no functioning primary/foreign key slot can have
+         * null values.
+         * beforeChange source: https://handsontable.com/docs/8.1.0/tutorial-using-callbacks.html#page-source-definition
+         * 
+         * @param {Array} grid_changes array of [row, column, old value (incl. undefined), new value (incl. null)].   
+         * @param {String} action can be CopyPaste.paste, ...
          */
         beforeChange: function (grid_changes, action) {
-          console.log("beforeChange")
-          if (!grid_changes) { // Is this ever the case?
-            return;
-          }          
+          // Ignore addition of new records to table.
+          if (action == 'add_row') return; 
+          if (!grid_changes) return; // Is this ever the case?
 
-          /* changes = array of [row, column, old value (incl. undefined), new 
-           * value (incl. null)].  NOTE: Cut & paste can include 2 or more 
+          /* TRICKY CASE: for tables depending on some other table for their 
+           * primary keys, grid_changes might involve "CopyPaste.paste" action
+           * onto other columns which cause new rows to be added at bottom 
+           * with empty values in read-only primary key fields. Need to extend
+           * change list in this case to include insertion of primary key(s).
+           */
+
+
+
+          console.log("beforeChange", grid_changes, action)
+
+          /* Cut & paste can include 2 or more 
            * changes in different columns on a row, or on multiple rows, so we 
            * need to validate 1 row at a time. Row_changes data structure enables
            * this.
-           * beforeChange source: https://handsontable.com/docs/8.1.0/tutorial-using-callbacks.html#page-source-definition
            */
-          let row_changes = self.getChanges(grid_changes); 
+          let row_changes = self.getRowChanges(grid_changes); 
 
           // Validate each row of changes
           for (const row in row_changes) {
@@ -826,9 +519,10 @@ class DataHarmonizer {
             }
 
             /* TEST FOR CHANGE TO UNIQUE_KEY
-             * If a class has a unique_key with >=1 slot then multiple changes 
-             * to key must be tested for duplicates as well as multiple
-             * possible impacts on dependent tables.
+             * Similar to a class's identifier slot(s), if a class has a 
+             * unique_key with >=1 slot then multiple changes to key must be
+             * tested for duplicates as well as multiple possible impacts on
+             * dependent tables.
              *
              * NOTE: user can be updating one field of a primary key; neets to 
              * be able to edit it in incomplete form.
@@ -880,54 +574,23 @@ class DataHarmonizer {
                   return false;
                 };
                 // Full changed unique_key here, and no duplicates, so on its
-                // own this is ok.  Below "dependent_name" test will prompt
-                // user if any implications to dependent records.
-
-              } // end if of keyVals construction.
-            } // end forEach of unique_keys
-
-            /* TEST FOR CHANGE TO DEPENDENT TABLE FOREIGN KEY(S)
-             * CASE 1: A shift from a complete foreign key to:
-             * A) Another complete combination.
-             *    1) duplicate?  
-             *       a) MERGE?
-             *    2) unique:
-             *       Prompt to change all underlying records?
-             * B) Deleted part
-             *    BLOCK.  Suggest deleting row.
-             * C) Cleared out.
-             *    Block - same as B).
-
-             * - changed key (part) -> prompt user to use "right click: Change primary key option"
-             *
-             * CASE 2: Shift from incomplete foreign key to complete one:
-             * - This doesn't necessitate any action except say for a refresh
-             *   of dependent table in case where we allow dependent to have 
-             *   "loose cannon" foreign key references.
-             * ISSUE IS THIS WOULD HAVE TO CASCADE KEY SENSITIVITY.
-             * WHERE DOES KEY SENSITIVITY RESPONSIBILITY END?
-             * BEST TO RECURSIVELY CHECK DEPENDENCIES
-
-
-            for (let dependent_name in self.context.crudGetChildren[self.template_name]) {
-              // dependent_class = ...
-              let [found, keyVals, change_log]
-               = self.context.crudCheckDependency(self, row, dependent_name, changes);
-              if (found && found == '') {
-                // found status: 
-                // 
-                // Here we have complete unique_keys entry to test for duplication
-                // There is a matching row so fail this action.
-                if (duplicate_row && duplicate_row != row) {
-                  alert(`Your key change on row ${parseInt(row)+1} would affect existing dependent records in the following tables! Do you want to continue with this change?  Check:\n\n ${change_log}`);
-                  return false;
-                };
+                // own this is ok.  Dependent table record test below will 
+                // prompt user if any implications to dependent records.
               }
             }
-                         */
-            // IF PRIMARY KEY CHANGE ACCEPTED ON MAIN TABLE AFTER PROMPT, THEN
-            // IMPLEMENT DATA CHANGE IN afterChange call...??? OR HERE TOO?
-            // ISSUE - 
+
+            // TEST FOR CHANGE TO DEPENDENT TABLE FOREIGN KEY(S)
+            // IF PRIMARY KEY CHANGE ACCEPTED BELOW AFTER PROMPT, THEN IMPLEMENT
+            // DATA CHANGE IN afterChange call...??? ISSUE is that technically 
+            // the underlying table data would be getting changed before the 
+            // change to this table's primary keys.
+            let warning = self.context.crudGetDependentUpdates(self.template_name, row);
+            // confirm(warning) presents user with report containing warning text.
+            if (!warning || confirm (warning+ "HI!")) {
+              if (warning)
+                // "True" parameter triggers updates
+                self.context.crudGetDependentUpdates(self.template_name, row, true);
+            } 
 
           } // end of row in row_change
 
@@ -941,7 +604,9 @@ class DataHarmonizer {
           if (triggered_changes) {
             grid_changes.push(...triggered_changes);
           }
+          //return false; will cancel this, yielding no changes.
         },
+
         // If a change is carried out in the grid but user doesn't
         // change selection, e.g. just edits 1 cell content, then
         // an afterSelection isn't triggered. THEREFORE we have to
@@ -949,12 +614,13 @@ class DataHarmonizer {
         afterChange: function (grid_changes, action) { // e.g. action=edit.
           // FUTURE: trigger only if change to primary key fields.
           console.log("afterChange", grid_changes, action)
-          //let row_changes = self.getChanges(grid_changes); 
+          //let row_changes = self.getRowChanges(grid_changes); 
           //Ignore if partial key left by change.  For full dependent key, 
           // Check changes against any dependent key slot. If match to at least
           // one slot, then trigger crudFilterDependentViews.
           self.context.crudFilterDependentViews(self.template_name);
         },
+
         afterSelection: (row, column, row2, column2) => {
           /* 
            * Test if user-focused row has changed.  If so, then possibly 
@@ -966,10 +632,8 @@ class DataHarmonizer {
            * FUTURE: 
            */
           if (self.current_selection[0] != row) {
-
             //console.log("afterSelection",self.template_name,self.context.relations)
             self.context.crudFilterDependentViews(self.template_name);
-
           };
 
           self.current_selection = [row, column, row2, column2];
@@ -1033,7 +697,7 @@ class DataHarmonizer {
    * tacitly ordered dict of row #s.
    * {[row]: {[slot_name]: {old_value: _, value: _} } }
    */
-  getChanges(grid_changes) { 
+  getRowChanges(grid_changes) { 
     let row_change = {}
 
     for (const change of grid_changes) {
@@ -1065,19 +729,20 @@ class DataHarmonizer {
     this.hot.updateSettings({ columns: currentColumns });
   }
 
-  /* Called via Footer.js, and also in right click menu DH contextMenu click 
+  /** Called via Footer.js, and also in right click menu DH contextMenu click 
    * row_above, row_below.
    * numRows is specified number of rows to add to current dh table. If table
    * has foreign key(s) to another table, the other table's focused row will
    * control foreign key values added to this table.
-   * row_where = 'insert_row_below' | 'insert_row_above'
+   * @param {String} row_where = 'insert_row_below' | 'insert_row_above'
   */
   addRows(row_where, numRows, startRowIndex = false) {
 
     numRows = parseInt(numRows); // Coming from form string input.
     // Get the starting row index where the new rows will be added
     if (startRowIndex === false)
-      startRowIndex = this.hot.countSourceRows(); // or .countSourceRows();???
+      startRowIndex = this.hot.countSourceRows();
+
 
     let parents = this.context.crudGetParents(this.template_name);
     // If this has no foreign key parent table(s) then go ahead and add x rows.
@@ -1090,7 +755,6 @@ class DataHarmonizer {
     // Here we have a dependent table that must locate values in its parent
     // table(s) for primary key values.
     let [required_selections, errors] = this.context.crudGetForeignKeyValues(parents);
-
     if (errors) {
       // Prompt user to select appropriate parent table row(s) first.
       $('#empty-parent-key-modal-info').html(errors);
@@ -1101,14 +765,20 @@ class DataHarmonizer {
     this.hot.alter(row_where, startRowIndex, numRows);
 
     // Populate new rows with selected foreign key value(s)
-    this.hot.batch(() => {
+    // PROBLEM: THIS IS TRIGGERING CHANGE test in subordinate tables
+    // GRDIIsolate setDataAtCell triggers AMRTest complaint
+    this.hot.batchRender(() => {
       for (let row = startRowIndex; row < startRowIndex + numRows; row++) {
         Object.entries(required_selections).forEach(([slot_name, value]) => {
           const col = this.getColumnIndexByFieldName(slot_name);
-          this.hot.setDataAtCell(row, col, value);
+          // "add_row" is critical so we can trap that in before_change event
+          // Otherwise a dependent view with "this change will impact ..." 
+          // message will be triggered.
+          this.hot.setDataAtCell(row, col, value.value, 'add_row');
         });
       };
     });
+
 
   }
 
@@ -1249,9 +919,9 @@ class DataHarmonizer {
     this.render();
   }
 
-  filterAll() {
+  filterAll(dh) {
     // Access the Handsontable instance's filter plugin
-    const filtersPlugin = this.hot.getPlugin('filters');
+    const filtersPlugin = dh.hot.getPlugin('filters');
 
     // Clear any existing filters
     filtersPlugin.clearConditions();
@@ -1264,15 +934,15 @@ class DataHarmonizer {
     filtersPlugin.filter();
   }
 
-  filterAllEmpty() {
-    const filtersPlugin = this.hot.getPlugin('filters');
+  filterAllEmpty(dh) {
+    const filtersPlugin = dh.hot.getPlugin('filters');
 
     // Clear any existing filters
     filtersPlugin.clearConditions();
 
     // Apply a filter condition to hide all empty rows
     filtersPlugin.addCondition((row /*col, value*/) => {
-      return !this.hot.getDataAtRow(row).every(isEmptyUnitVal);
+      return !dh.hot.getDataAtRow(row).every(isEmptyUnitVal);
     });
 
     filtersPlugin.filter();
@@ -1532,28 +1202,6 @@ class DataHarmonizer {
     this.hot.scrollViewportTo(row, column);
   }
 
-  // TODO
-  // /**
-  //  * Run void function behind loading screen.
-  //  * Adds function to end of call queue. Does not handle functions with return
-  //  * vals, unless the return value is a promise. Even then, it only waits for the
-  //  * promise to resolve, and does not actually do anything with the value
-  //  * returned from the promise.
-  //  * @param {function} fn - Void function to run.
-  //  * @param {Array} [args=[]] - Arguments for function to run.
-  //  */
-  async runBehindLoadingScreen(fn, args = []) {
-    await $('#loading-screen').show('fast', 'swing');
-    await wait(200); // Enough of a visual cue that something is happening
-    if (args.length) {
-      await fn.apply(this, args);
-    } else {
-      await fn.apply(this);
-    }
-    await $('#loading-screen').hide();
-    return;
-  }
-
   /***************************** PRIVATE functions *************************/
 
   /**
@@ -1682,7 +1330,7 @@ class DataHarmonizer {
             specifiedHeaderRow - 1
           );
           $('#specify-headers-modal').modal('hide');
-          self.runBehindLoadingScreen(() => {
+          self.context.runBehindLoadingScreen(() => {
             self.hot.loadData(
               self.matrixFieldChangeRules(mappedMatrixObj.matrix.slice(2))
             );
@@ -1962,23 +1610,23 @@ class DataHarmonizer {
    */
   enableMultiSelection() {
     const self = this;
+    const dh = this.hot;
     this.hot.updateSettings({
       afterBeginEditing: function (row, col) {
 
         if (self.slots[col].flatVocabulary && self.slots[col].multivalued === true) {
-          const value = this.getDataAtCell(row, col);
+          const value = dh.getDataAtCell(row, col);
           const selections = parseMultivaluedValue(value);
           const formattedValue = formatMultivaluedValue(selections);
 
           // Cleanup of empty values that can occur with leading/trailing or double ";"
           if (value !== formattedValue) {
-            this.setDataAtCell(row, col, formattedValue, 'thisChange');
+            dh.setDataAtCell(row, col, formattedValue, 'thisChange');
           }
-
           let content = '';
-          if (this.slots[col].sources) {
-            this.slots[col].sources.forEach((source) => {
-              Object.values(this.slots[col].permissible_values[source]).forEach(
+          if (self.slots[col].sources) {
+            self.slots[col].sources.forEach((source) => {
+              Object.values(self.slots[col].permissible_values[source]).forEach(
                 (permissible_value) => {
                   const field = permissible_value.text;
                   const field_trim = field.trim();
@@ -1995,7 +1643,7 @@ class DataHarmonizer {
 
           $('#field-description-text').html(
             `<span data-i18n="${''}">
-            ${this.slots[col].title}
+            ${self.slots[col].title}
               </span>
             <select multiple class="multiselect" rows="15">${content}</select>`
           );
@@ -2020,7 +1668,7 @@ class DataHarmonizer {
               let newValCsv = formatMultivaluedValue(
                 $('#field-description-text .multiselect').val()
               );
-              self.setDataAtCell(row, col, newValCsv, 'thisChange');
+              dh.setDataAtCell(row, col, newValCsv, 'thisChange');
             });
           // Saves users a click:
           $('#field-description-text .multiselect')[0].selectize.focus();
@@ -2059,82 +1707,6 @@ class DataHarmonizer {
     }
   }
 
-  /**
-   * Recursively flatten vocabulary into an array of strings, with each
-   * string's level of depth in the vocabulary being indicated by leading
-   * spaces.
-   * FUTURE possible functionality:
-   * Both validation and display of picklist item becomes conditional on
-   * other picklist item if "depends on" indicated in picklist item/branch.
-   * @param {Object} vocabulary See `vocabulary` fields in SCHEMA.
-   * @return {Array<String>} Flattened vocabulary.
-   */
-  stringifyNestedVocabulary(vocab_list) {
-    let ret = [];
-    let stack = [];
-    for (const pointer in vocab_list) {
-      let choice = vocab_list[pointer];
-      let level = 0;
-      if ('is_a' in choice) {
-        level = stack.indexOf(choice.is_a) + 1;
-        stack.splice(level + 1, 1000, choice.text);
-      } else {
-        stack = [choice.text];
-      }
-
-      this.setExportField(choice, false);
-
-      ret.push('  '.repeat(level) + choice.text);
-    }
-    return ret;
-  }
-
-  setExportField(field, as_field) {
-    if (field.exact_mappings) {
-      field.exportField = {};
-      for (let item of field.exact_mappings) {
-        let ptr = item.indexOf(':');
-        if (ptr != -1) {
-          const prefix = item.substr(0, ptr);
-          if (!(prefix in field.exportField)) {
-            field.exportField[prefix] = [];
-          }
-
-          const mappings = item.substr(ptr + 1);
-          for (let mapping of mappings.split(';')) {
-            mapping = mapping.trim();
-            const conversion = {};
-            //A colon alone means to map value to empty string
-            if (mapping == ':') {
-              conversion.value = '';
-            }
-            //colon with contents = field & value
-            else {
-              mapping = decodeURIComponent(mapping);
-              if (mapping.indexOf(':') != -1) {
-                const binding = mapping.split(':');
-                binding[0] = binding[0].trim();
-                binding[1] = binding[1].trim();
-                if (binding[0] > '') {
-                  conversion.field = binding[0];
-                }
-                if (binding[1] > '') {
-                  conversion.value = binding[1];
-                }
-              }
-              //No colon means its just field or value
-              else if (as_field == true) {
-                conversion.field = mapping;
-              } else {
-                conversion.value = mapping;
-              }
-            }
-            field.exportField[prefix].push(conversion);
-          }
-        }
-      }
-    }
-  }
 
   /**
    * Get an HTML string that describes a field, its examples etc. for display
